target variable to float?

+++ src/randomForest_model.py
@@ -20,7 +20,7 @@ def rfModelTest():
 
 def rfModel(dirs, exprfile, pheno, genes, cvs):
     print("random forest")
-    ys = [int(yi) for yi in gzip.open(dirs + pheno,'rt').read().strip().split('\n')]
+    ys = [float(yi) for yi in gzip.open(dirs + pheno,'rt').read().strip().split('\n')]
     inputs = gzip.open(dirs + exprfile, 'rt').read().strip().split("\n")
     inputHeader = inputs.pop(0)
     scoreList = []
@@ -45,7 +45,7 @@ def rfModel(dirs, exprfile, pheno, genes, cvs):
 
 def rfModelSetScores(dirs, inputs, pheno, genes, cvs):
     print("random forest")
-    ys = [int(yi) for yi in gzip.open(dirs + pheno,'rt').read().strip().split('\n')]
+    ys = [float(yi) for yi in gzip.open(dirs + pheno,'rt').read().strip().split('\n')]
     # use all reported gene sets for prediction
     xs = np.array([ np.array(x) for x in inputs ])
     clf = RandomForestClassifier(max_depth=5, n_estimators=100)
